pyproject.toml
[project]
name = "the-complete-agentic-ai-engineering"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.57.1",
    "google-api-python-client>=2.176.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "google-generativeai>=0.8.5",
    "gradio>=5.37.0",
    "ipywidgets>=8.1.7",
    "notebook>=7.4.4",
    "openai>=1.97.0",
    "openai-agents>=0.2.10",
    "pypdf>=5.8.0",
    "python-decouple>=3.8",
    "python-dotenv>=1.1.1",
    "sendgrid>=6.12.4",
]
